+++ Sources/IBLinterCore/Models/Documents/StoryboardDocument.swift
@@ -1,11 +1,13 @@
 //
-//  swift
+//  StoryboardDocument.swift
 //  IBLinterCore
 //
-//  Created by SaitoYuta on 2017/12/05.
+//  Created by SaitoYuta on 3/11/18.
 //
 
 import SWXMLHash
+
+// MARK: - StoryboardDocument
 
 public struct StoryboardDocument: XMLDecodable {
     public let type: String
@@ -41,37 +43,7 @@ public struct StoryboardDocument: XMLDecodable {
     }
 }
 
-public struct XibDocument: XMLDecodable {
-    public let type: String
-    public let version: String
-    public let toolsVersion: String
-    public let targetRuntime: String
-    public let propertyAccessControl: String?
-    public let useAutolayout: Bool?
-    public let useTraitCollections: Bool?
-    public let useSafeAreas: Bool?
-    public let colorMatched: Bool?
-    public let device: Device?
-    public let views: [AnyView]?
-    public let placeholders: [Placeholder]?
-
-    static func decode(_ xml: XMLIndexer) throws -> XibDocument {
-        return XibDocument.init(
-            type:                  try xml.attributeValue(of: "type"),
-            version:               try xml.attributeValue(of: "version"),
-            toolsVersion:          try xml.attributeValue(of: "toolsVersion"),
-            targetRuntime:         try xml.attributeValue(of: "targetRuntime"),
-            propertyAccessControl: xml.attributeValue(of: "propertyAccessControl"),
-            useAutolayout:         xml.attributeValue(of: "useAutolayout"),
-            useTraitCollections:   xml.attributeValue(of: "useTraitCollections"),
-            useSafeAreas:          xml.attributeValue(of: "useSafeAreas"),
-            colorMatched:          xml.attributeValue(of: "colorMatched"),
-            device:                xml.byKey("device").flatMap(decodeValue),
-            views:                 try xml.byKey("objects")?.children.flatMap(AnyView.decode),
-            placeholders:          xml.byKey("objects")?.byKey("placeholder")?.all.flatMap(decodeValue)
-        )
-    }
-}
+// MARK: - Device
 
 public struct Device: XMLDecodable {
     public let id: String
@@ -87,6 +59,8 @@ public struct Device: XMLDecodable {
     }
 }
 
+// MARK: - Scene
+
 public struct Scene: XMLDecodable {
     public let id: String
     public let viewController: AnyViewController?
@@ -98,6 +72,8 @@ public struct Scene: XMLDecodable {
         )
     }
 }
+
+// MARK: - Placeholder
 
 public struct Placeholder: XMLDecodable {
     public let id: String
@@ -116,54 +92,3 @@ public struct Placeholder: XMLDecodable {
         )
     }
 }
-
-public struct ViewControllerLayoutGuide: XMLDecodable {
-    public let id: String
-    public let type: String
-
-    static func decode(_ xml: XMLIndexer) throws -> ViewControllerLayoutGuide {
-        return try ViewControllerLayoutGuide.init(
-            id: xml.attributeValue(of: "id"),
-            type: xml.attributeValue(of: "type")
-        )
-    }
-}
-
-public struct LayoutGuide: XMLDecodable {
-    public let key: String
-    public let id: String
-
-    static func decode(_ xml: XMLIndexer) throws -> LayoutGuide {
-        return try LayoutGuide.init(
-            key: xml.attributeValue(of: "key"),
-            id: xml.attributeValue(of: "id")
-        )
-    }
-}
-
-public enum IBError: Swift.Error, CustomStringConvertible {
-    case elementNotFound
-    case unsupportedViewClass(String)
-    case unsupportedViewControllerClass(String)
-    case unsupportedConstraint(String)
-    case unsupportedTableViewDataMode(String)
-    case unsupportedColorSpace(String)
-
-    public var description: String {
-        switch self {
-        case .elementNotFound:
-            return "element not found"
-        case .unsupportedViewClass(let name):
-            return "unsupported view class '\(name)'"
-        case .unsupportedViewControllerClass(let name):
-            return "unsupported viewController class '\(name)'"
-        case .unsupportedConstraint(let body):
-            return "unsupported constraint type '\(body)'"
-        case .unsupportedTableViewDataMode(let name):
-            return "unsupported dataMode '\(name)'"
-        case .unsupportedColorSpace(let colorSpace):
-            return "unsupported color space '\(colorSpace)'"
-        }
-    }
-}
-
