+++ Sources/IBLinterCore/Models/Files/XibFile.swift
@@ -9,10 +9,8 @@ import SWXMLHash
 import Foundation
 
 public class XibFile: InterfaceBuilderFile {
+
     public var pathString: String
-    public var fileName: String {
-        return pathString.components(separatedBy: "/").last!
-    }
 
     public let document: XibDocument
