+++ Sources/IBLinterCore/Models/Views/ScrollViews/ScrollView.swift
@@ -1,0 +1,44 @@
+//
+//  ScrollView.swift
+//  IBLinterCore
+//
+//  Created by SaitoYuta on 3/11/18.
+//
+
+import SWXMLHash
+
+public struct ScrollView: XMLDecodable, ViewProtocol {
+    public let id: String
+    public let elementClass: String = "UIScrollView"
+
+    public let autoresizingMask: AutoresizingMask?
+    public let clipsSubviews: Bool?
+    public let constraints: [Constraint]?
+    public let contentMode: String?
+    public let customClass: String?
+    public let customModule: String?
+    public let isMisplaced: Bool?
+    public let opaque: Bool?
+    public let rect: Rect
+    public let subviews: [AnyView]?
+    public let translatesAutoresizingMaskIntoConstraints: Bool?
+    public let userInteractionEnabled: Bool?
+
+    static func decode(_ xml: XMLIndexer) throws -> ScrollView {
+        return ScrollView.init(
+            id:                                        try xml.attributeValue(of: "id"),
+            autoresizingMask:                          xml.byKey("autoresizingMask").flatMap(decodeValue),
+            clipsSubviews:                             xml.attributeValue(of: "clipsSubviews"),
+            constraints:                               try xml.byKey("constraints")?.byKey("constraint")?.all.flatMap(Constraint.decode),
+            contentMode:                               xml.attributeValue(of: "contentMode"),
+            customClass:                               xml.attributeValue(of: "customClass"),
+            customModule:                              xml.attributeValue(of: "customModule"),
+            isMisplaced:                               xml.attributeValue(of: "misplaced"),
+            opaque:                                    xml.attributeValue(of: "opaque"),
+            rect:                                      try decodeValue(xml.byKey("rect")),
+            subviews:                                  try xml.byKey("subviews")?.children.flatMap(AnyView.decode),
+            translatesAutoresizingMaskIntoConstraints: xml.attributeValue(of: "translatesAutoresizingMaskIntoConstraints"),
+            userInteractionEnabled:                    xml.attributeValue(of: "userInteractionEnabled")
+        )
+    }
+}
