+++ Sources/IBLinterCore/Models/Views/StackView.swift
@@ -1,0 +1,48 @@
+//
+//  StackView.swift
+//  IBLinterCore
+//
+//  Created by SaitoYuta on 3/11/18.
+//
+
+import SWXMLHash
+
+public struct StackView: XMLDecodable, ViewProtocol {
+    public let id: String
+    public let elementClass: String = "UIStackView"
+
+    public let alignment: String?
+    public let autoresizingMask: AutoresizingMask?
+    public let axis: String?
+    public let clipsSubviews: Bool?
+    public let constraints: [Constraint]?
+    public let contentMode: String?
+    public let customClass: String?
+    public let customModule: String?
+    public let isMisplaced: Bool?
+    public let opaque: Bool?
+    public let rect: Rect
+    public let subviews: [AnyView]?
+    public let translatesAutoresizingMaskIntoConstraints: Bool?
+    public let userInteractionEnabled: Bool?
+
+    static func decode(_ xml: XMLIndexer) throws -> StackView {
+        return StackView.init(
+            id:                                        try xml.attributeValue(of: "id"),
+            alignment:                                 xml.attributeValue(of: "alignment"),
+            autoresizingMask:                          xml.byKey("autoresizingMask").flatMap(decodeValue),
+            axis:                                      xml.attributeValue(of: "axis"),
+            clipsSubviews:                             xml.attributeValue(of: "clipsSubviews"),
+            constraints:                               try xml.byKey("constraints")?.byKey("constraint")?.all.flatMap(Constraint.decode),
+            contentMode:                               xml.attributeValue(of: "contentMode"),
+            customClass:                               xml.attributeValue(of: "customClass"),
+            customModule:                              xml.attributeValue(of: "customModule"),
+            isMisplaced:                               xml.attributeValue(of: "misplaced"),
+            opaque:                                    xml.attributeValue(of: "opaque"),
+            rect:                                      try decodeValue(xml.byKey("rect")),
+            subviews:                                  try xml.byKey("subviews")?.children.flatMap(AnyView.decode),
+            translatesAutoresizingMaskIntoConstraints: xml.attributeValue(of: "translatesAutoresizingMaskIntoConstraints"),
+            userInteractionEnabled:                    xml.attributeValue(of: "userInteractionEnabled")
+        )
+    }
+}
