+++ Sources/IBLinterCore/Models/Views/Toolbar.swift
@@ -1,0 +1,63 @@
+//
+//  Toolbar.swift
+//  IBLinterCore
+//
+//  Created by SaitoYuta on 3/11/18.
+//
+
+import SWXMLHash
+
+public struct Toolbar: XMLDecodable, ViewProtocol {
+    public let id: String
+    public let elementClass: String = "UIToolbar"
+
+    public let autoresizingMask: AutoresizingMask?
+    public let clipsSubviews: Bool?
+    public let constraints: [Constraint]?
+    public let contentMode: String?
+    public let customClass: String?
+    public let customModule: String?
+    public let items: [BarButtonItem]?
+    public let isMisplaced: Bool?
+    public let opaque: Bool?
+    public let rect: Rect
+    public let subviews: [AnyView]?
+    public let translatesAutoresizingMaskIntoConstraints: Bool?
+    public let userInteractionEnabled: Bool?
+
+
+    public struct BarButtonItem: XMLDecodable {
+        public let id: String
+        public let style: String?
+        public let systemItem: String?
+        public let title: String?
+
+        static func decode(_ xml: XMLIndexer) throws -> Toolbar.BarButtonItem {
+            return BarButtonItem.init(
+                id:         try xml.attributeValue(of: "id"),
+                style:      xml.attributeValue(of: "style"),
+                systemItem: xml.attributeValue(of: "systemItem"),
+                title:      xml.attributeValue(of: "title")
+            )
+        }
+    }
+
+    static func decode(_ xml: XMLIndexer) throws -> Toolbar {
+        return Toolbar.init(
+            id:                                        try xml.attributeValue(of: "id"),
+            autoresizingMask:                          xml.byKey("autoresizingMask").flatMap(decodeValue),
+            clipsSubviews:                             xml.attributeValue(of: "clipsSubviews"),
+            constraints:                               try xml.byKey("constraints")?.byKey("constraint")?.all.flatMap(Constraint.decode),
+            contentMode:                               xml.attributeValue(of: "contentMode"),
+            customClass:                               xml.attributeValue(of: "customClass"),
+            customModule:                              xml.attributeValue(of: "customModule"),
+            items:                                     try xml.byKey("items")?.byKey("barButtonItem")?.all.flatMap(BarButtonItem.decode),
+            isMisplaced:                               xml.attributeValue(of: "misplaced"),
+            opaque:                                    xml.attributeValue(of: "opaque"),
+            rect:                                      try decodeValue(xml.byKey("rect")),
+            subviews:                                  try xml.byKey("subviews")?.children.flatMap(AnyView.decode),
+            translatesAutoresizingMaskIntoConstraints: xml.attributeValue(of: "translatesAutoresizingMaskIntoConstraints"),
+            userInteractionEnabled:                    xml.attributeValue(of: "userInteractionEnabled")
+        )
+    }
+}
