+++ Sources/IBLinterCore/Models/Views/View.swift
@@ -1,0 +1,61 @@
+//
+//  View.swift
+//  IBLinterCore
+//
+//  Created by SaitoYuta on 3/11/18.
+//
+
+import SWXMLHash
+
+public struct View: XMLDecodable, ViewProtocol {
+
+    public let id: String
+    public let elementClass: String = "UIView"
+
+    public let autoresizingMask: AutoresizingMask?
+    public let clipsSubviews: Bool?
+    public let constraints: [Constraint]?
+    public let contentMode: String?
+    public let customClass: String?
+    public let customModule: String?
+    public let isMisplaced: Bool?
+    public let opaque: Bool?
+    public let rect: Rect
+    public let subviews: [AnyView]?
+    public let translatesAutoresizingMaskIntoConstraints: Bool?
+    public let userInteractionEnabled: Bool?
+    public let viewLayoutGuide: LayoutGuide?
+
+    static func decode(_ xml: XMLIndexer) throws -> View {
+        return View.init(
+            id:                                        try xml.attributeValue(of: "id"),
+            autoresizingMask:                          xml.byKey("autoresizingMask").flatMap(decodeValue),
+            clipsSubviews:                             xml.attributeValue(of: "clipsSubviews"),
+            constraints:                               try xml.byKey("constraints")?.byKey("constraint")?.all.flatMap(Constraint.decode),
+            contentMode:                               xml.attributeValue(of: "contentMode"),
+            customClass:                               xml.attributeValue(of: "customClass"),
+            customModule:                              xml.attributeValue(of: "customModule"),
+            isMisplaced:                               xml.attributeValue(of: "misplaced"),
+            opaque:                                    xml.attributeValue(of: "opaque"),
+            rect:                                      try decodeValue(xml.byKey("rect")),
+            subviews:                                  try xml.byKey("subviews")?.children.flatMap(AnyView.decode),
+            translatesAutoresizingMaskIntoConstraints: xml.attributeValue(of: "translatesAutoresizingMaskIntoConstraints"),
+            userInteractionEnabled:                    xml.attributeValue(of: "userInteractionEnabled"),
+            viewLayoutGuide:                           xml.byKey("viewLayoutGuide").flatMap(decodeValue)
+        )
+    }
+}
+
+// MARK: - LayoutGuide
+
+public struct LayoutGuide: XMLDecodable {
+    public let key: String
+    public let id: String
+
+    static func decode(_ xml: XMLIndexer) throws -> LayoutGuide {
+        return try LayoutGuide.init(
+            key: xml.attributeValue(of: "key"),
+            id: xml.attributeValue(of: "id")
+        )
+    }
+}
